Fix directional balancing
If using field relative, forward and backwards on the field are static which doesn't work based on who you approach the charging pad. If approaching from robot forward, pitch will be positive but this would only work on one side of the charging station.

For balancing, robot forward and backwards need to be used.